New setCapacity expects Double, not Int. Fixed in order to be consistent with Writer.
git-svn-id: file:///srv/local/data/cleanRepo/trunk@6511 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/org/matsim/core/facilities/FacilitiesReaderMatsimV1.java
@@ -102,7 +102,7 @@ public class FacilitiesReaderMatsimV1 extends MatsimXmlParser {
 	}
 	
 	private void startCapacity(final Attributes atts) {
-		int cap = Integer.parseInt(atts.getValue("value"));
+		double cap = Double.parseDouble(atts.getValue("value"));
 		this.curractivity.setCapacity(cap);
 	}
